Part 1 of GPS timing refactor.
Removes most usage of DateTime and substitutes it with usage of
GPS Epoch Time.  Stubs out a couple of methods that need to be
completed.  They are in dateTime.c.  Look for GPS_DATE_TIME_STUB.
The methods are documented in dateTime.h in case their meaning isn't
clear.

TODO:
  * Implement the stubbed methods
  * Fix up the Unit tests that broke as a result
  * Write new tests for the stubbed methods.
  * Profit!