Add in GPS heading support
GPS heading allows us to see what direciton a vehicle is truely moving
in.  This can in turn be used for any number of applications.  The
implementation is simple in that it compares a previous good point to
its current point to determine a current heading.  It is limited by the
rate at which the GPS unit updates, and currently only works above 10MPH.
0 indicates that no value could be read.  True north is represented by
a value of 360.