Simple fix to only calculate GPS heading if speed > 10 MPH
This is to avoid issues seen when drivers move too slow.  It is
really only a temporary fix and is generally poor.  A better fix
is to record GPS history up to 1 second back and take the point
that is closest to the current point time wise but meets a certain
minimum distance.  This would allow us to get the maximum amount of
accuracy with the variable GPS sample rate that we support.