Initial addition of slip_angle code
This change adds in all the code needed to calculate a rudimentary
slip angle based on gps heading and the integral of the yaw sender
within the vehicle.

Note that this is not the final implementation.  Rather this will
only get us a rough estimate and may be buggy.  But its a start.