Add the initial unit tests for slip angle calculations
This change adds the initial unit tests for slip_angle
to ensure that the methods are working as expected.  It does
not cover the check for if we are going straight or not as I
have not determined whether or not that code will remain.  Until
I have a better idea on the matter it shall remain testless.